Revert "Revert "Datetype fix""
This reverts commit 4a7e8453b1f02d82fbe46435a3980c5f6ec41bf6.

+++ REST_API/REST_API/Models/Api/Attachments/AttachmentRequest.cs
@@ -8,6 +8,6 @@ namespace REST_API.Models.Api.Attachments
     public class AttachmentRequest
     {
         public uint Id_Attachment { get; set; }
-        public uint Id_Message { get; set; }
+        public ulong Id_Message { get; set; }
     }
 }
